requirements.txt
aiofiles==0.6.0
aniso8601==7.0.0
appdirs==1.4.4
astroid==2.3.3
async-exit-stack==1.0.1
async-generator==1.10
bandit==1.7.0
beautifulsoup4==4.9.3
black==21.5b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.47.1
gitdb==4.0.7
GitPython==3.1.17
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
gunicorn==20.0.4
h11==0.9.0
httptools==0.0.13
idna==2.10
isort==4.3.21
itsdangerous==2.0.0
Jinja2==3.0.0
lazy-object-proxy==1.4.3
loguru==0.5.3
MarkupSafe==2.0.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pbr==5.6.0
promise==2.3
psycopg2==2.8.6
pydantic==1.4
pylint==2.4.4
python-igraph==0.9.1
python-multipart==0.0.5
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
Rx==1.6.1
six==1.14.0
smmap==4.0.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
starlette==0.12.9
stevedore==3.3.0
texttable==1.6.3
toml==0.10.2
ujson==4.0.2
Unidecode==1.2.0
urllib3==1.26.4
uvicorn==0.11.2
uvloop==0.15.2
websockets==8.1
wikipedia==1.4.0
wrapt==1.11.2
